Access rules via friend statement instead of #define private public
Added a friend statement in class EClientSocketBase and
EPosixClientSocket to replace the trick with teh #define. This prevent
some compilers will complain about different access rules in the
standard c++ libraries.
Because the next version of the IB implementation of this API is so
different, changing these files does not create a maintenance burden